revert the change of outside PortEntity and add internal port entity then rename package name to dataplance
